Added __task specifier
This specifier informs the compiler that the function is a task. This
means that the function can disregard the normal calling conventions and
only preserve the LR register. In the current implementation we don't
need to preserve the LR register but we'll need it in the future when
tasks can be terminated.