Systick will now continue running during syscalls
The time spent context switching and in the kernel is now considered
when decrementing the amount of sleep time tasks have left. This fixes
the issue where sleep() calls would last much longer in applications
with a huge amount of syscalls. For example, a system with 2 tasks that
yield() and 1 task that sleeps would result in the sleep() lasting 4x
too long.